+++ Changelog.md
@@ -4,7 +4,7 @@
 
 - Add flag wybin search user
 - Mantain staff flags when adding a new staff member.
-
+- Add flag staff registration
 
 ## Version 3.4.5